feat: simpler op definitions
