+++ src/service.ts
@@ -1,8 +1,8 @@
 import { EndpointEntityMap } from './endpoint.js';
-import { Operation, OpFactoryOptions, OpFunction, paramsFromOptions } from './ops.js';
-import { LeaveRequest, Terminal } from './rotacloud.js';
+import { LeaveRequest, Terminal } from './interfaces/index.js';
 import { Options } from './fetchv2.js';
 import { LaunchTerminal } from './interfaces/launch-terminal.interface.js';
+import { Op, Operation, OpFactoryOptions, paramsFromOptions } from './ops.js';
 
 export type RequirementsOf<T, K extends keyof T> = Required<Pick<T, K>> & Partial<T>;
 
@@ -13,7 +13,7 @@ export type ServiceSpecification = {
    * Operations unique to the service
    * Can be used to override operations listed in {@see ServiceSpecification['operations']}
    */
-  customOperations?: Record<string, (opts: OpFactoryOptions) => (...args: any[]) => any>;
+  customOperations?: Record<string, Op<any>>;
 } & (
   | {
       /** URL of the endpoint */
@@ -104,31 +104,25 @@ export const SERVICES = {
     endpointVersion: 'v1',
     operations: ['get', 'list', 'delete', 'create'],
     customOperations: {
-      create:
-        ({
-          client,
-          request,
-          service,
-        }: OpFactoryOptions): OpFunction<LeaveRequest, RequirementsOf<LeaveRequest, 'user'>> =>
-        async <T = LeaveRequest>(entity: RequirementsOf<LeaveRequest, 'user'>, opts?: Options<T>) => {
-          const modifiedRequest = {
-            ...request,
-            headers: {
-              ...request.headers,
-              Account: undefined,
-              User: entity.user,
-            },
-            params: {
-              ...request.params,
-              ...paramsFromOptions(opts ?? {}),
-            },
-          };
-          // assert(request.headers !== undefined, 'Invalid create leave request');
-          const res = await client.post<T>(service.endpoint, entity, modifiedRequest);
-
-          if (opts?.rawResponse) return res;
-          return res.data;
-        },
+      // create: (
+      //   { request, service }: OpFactoryOptions,
+      //   entity: RequirementsOf<LeaveRequest, 'user'>,
+      //   opts?: Options<LeaveRequest>,
+      // ) => ({
+      //   ...request,
+      //   method: 'POST',
+      //   url: service.endpoint,
+      //   data: entity,
+      //   headers: {
+      //     ...request.headers,
+      //     Account: undefined,
+      //     User: entity.user,
+      //   },
+      //   params: {
+      //     ...request.params,
+      //     ...paramsFromOptions(opts ?? {}),
+      //   },
+      // }),
     },
   },
   location: {
@@ -156,49 +150,36 @@ export const SERVICES = {
     endpointVersion: 'v1',
     operations: ['create', 'get', 'list', 'listAll', 'update', 'delete'],
     customOperations: {
-      acknowledge:
-        ({ client, request }): OpFunction<void, number[]> =>
-        async <T = void, P = number[]>(shiftIds: P, opts?: Options<T>) => {
-          const modifiedRequest = {
-            ...request,
-            params: {
-              ...request.params,
-              ...paramsFromOptions(opts ?? {}),
-            },
-          };
-          const res = await client.post('/shifts_acknowledged', { shifts: shiftIds }, modifiedRequest);
-          if (opts?.rawResponse) return res;
-          return res.data;
-        },
-      publish:
-        ({ client, request }): OpFunction<void, number[]> =>
-        async <T = void, P = number[]>(shiftIds: P, opts?: Options<T>) => {
-          const modifiedRequest = {
-            ...request,
-            params: {
-              ...request.params,
-              ...paramsFromOptions(opts ?? {}),
-            },
-          };
-          const res = await client.post('/shifts_published', { shifts: shiftIds }, modifiedRequest);
-          if (opts?.rawResponse) return res;
-          return res.data;
-        },
-      unpublish:
-        ({ client, request }): OpFunction<void, number[]> =>
-        async <T = void, P = number[]>(shiftIds: P, opts?: Options<T>) => {
-          const modifiedRequest = {
-            ...request,
-            params: {
-              ...request.params,
-              ...paramsFromOptions(opts ?? {}),
-            },
-            data: { shifts: shiftIds },
-          };
-          const res = await client.delete('/shifts_published', modifiedRequest);
-          if (opts?.rawResponse) return res;
-          return res.data;
-        },
+      // acknowledge: ({ request }, shiftIds: number[], opts?: Options<void>) => ({
+      //   ...request,
+      //   method: 'POST',
+      //   url: '/shifts_acknowledged',
+      //   data: { shifts: shiftIds },
+      //   params: {
+      //     ...request.params,
+      //     ...paramsFromOptions(opts ?? {}),
+      //   },
+      // }),
+      // publish: ({ request }, shiftIds: number[], opts?: Options<void>) => ({
+      //   ...request,
+      //   method: 'POST',
+      //   url: '/shifts_published',
+      //   data: { shifts: shiftIds },
+      //   params: {
+      //     ...request.params,
+      //     ...paramsFromOptions(opts ?? {}),
+      //   },
+      // }),
+      // unpublish: ({ request }, shiftIds: number[], opts?: Options<void>) => ({
+      //   ...request,
+      //   method: 'DELETE',
+      //   url: '/shifts_published',
+      //   data: { shifts: shiftIds },
+      //   params: {
+      //     ...request.params,
+      //     ...paramsFromOptions(opts ?? {}),
+      //   },
+      // }),
     },
   },
   terminal: {
@@ -206,20 +187,15 @@ export const SERVICES = {
     endpointVersion: 'v1',
     operations: ['create', 'get', 'update', 'list', 'listAll'],
     customOperations: {
-      close:
-        ({ client, request, service }): OpFunction<void, number> =>
-        async <T = void, P = number>(id: P, opts?: Options<T>) => {
-          const modifiedRequest = {
-            ...request,
-            params: {
-              ...request.params,
-              ...paramsFromOptions(opts ?? {}),
-            },
-          };
-          const res = await client.delete(`${service.endpoint}/${id}`, modifiedRequest);
-          if (opts?.rawResponse) return res;
-          return res.data;
-        },
+      // close: ({ request, service }, id: number, opts?: Options<void>) => ({
+      //   ...request,
+      //   method: 'DELETE',
+      //   url: `/${service.endpoint}/${id}`,
+      //   params: {
+      //     ...request.params,
+      //     ...paramsFromOptions(opts ?? {}),
+      //   },
+      // }),
     },
   },
   terminalActive: {
@@ -227,48 +203,33 @@ export const SERVICES = {
     endpointVersion: 'v1',
     operations: ['list', 'listAll'],
     customOperations: {
-      launch:
-        ({ client, request, service }): OpFunction<Terminal, LaunchTerminal> =>
-        async <T = Terminal, P = LaunchTerminal>(id: P, opts?: Options<T>) => {
-          const modifiedRequest = {
-            ...request,
-            params: {
-              ...request.params,
-              ...paramsFromOptions(opts ?? {}),
-            },
-          };
-          const res = await client.delete(`${service.endpoint}/${id}`, modifiedRequest);
-          if (opts?.rawResponse) return res;
-          return res.data;
-        },
-      ping:
-        ({ client, request, service }): OpFunction<void, { id: number; action: string; device: string }> =>
-        async <T = void, P = { id: number; action: string; device: string }>(id: P, opts?: Options<T>) => {
-          const modifiedRequest = {
-            ...request,
-            params: {
-              ...request.params,
-              ...paramsFromOptions(opts ?? {}),
-            },
-          };
-          const res = await client.delete(`${service.endpoint}/${id}`, modifiedRequest);
-          if (opts?.rawResponse) return res;
-          return res.data;
-        },
-      close:
-        ({ client, request, service }): OpFunction<void, number> =>
-        async <T = void, P = number>(id: P, opts?: Options<T>) => {
-          const modifiedRequest = {
-            ...request,
-            params: {
-              ...request.params,
-              ...paramsFromOptions(opts ?? {}),
-            },
-          };
-          const res = await client.delete(`${service.endpoint}/${id}`, modifiedRequest);
-          if (opts?.rawResponse) return res;
-          return res.data;
-        },
+      // launch: ({ request, service }, id: LaunchTerminal, opts?: Options<Terminal>) => ({
+      //   ...request,
+      //   method: 'DELETE',
+      //   url: `/${service.endpoint}/${id}`,
+      //   params: {
+      //     ...request.params,
+      //     ...paramsFromOptions(opts ?? {}),
+      //   },
+      // }),
+      // ping: ({ request, service }, id: { id: number; action: string; device: string }, opts?: Options<void>) => ({
+      //   ...request,
+      //   method: 'DELETE',
+      //   url: `${service.endpoint}/${id}`,
+      //   params: {
+      //     ...request.params,
+      //     ...paramsFromOptions(opts ?? {}),
+      //   },
+      // }),
+      // close: ({ request, service }, id: number, opts?: Options<void>) => ({
+      //   ...request,
+      //   method: 'DELETE',
+      //   url: `${service.endpoint}/${id}`,
+      //   params: {
+      //     ...request.params,
+      //     ...paramsFromOptions(opts ?? {}),
+      //   },
+      // }),
     },
   },
   timeZone: {
